services/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.9
